+++ mayan/apps/document_parsing/tests/literals.py
@@ -1,3 +1,3 @@
 from __future__ import unicode_literals
 
-TEST_PARSING_INDEX_NODE_TEMPLATE = '{% if "sample" in document.latest_version.content|join:" "|lower %}sample{% endif %}'
+TEST_PARSING_INDEX_NODE_TEMPLATE = '{% if "sample" in document.content.lower() %}sample{% endif %}'