+++ HISTORY.rst
@@ -83,6 +83,8 @@
   and converteed before serializing them.
 - Add the 'ocr_content' attribute to documents to allow access
   to a document's OCR content for indexing and other purposes.
+- Add the 'content' attribute to documents to allow access
+  to a document's parsed content for indexing and other purposes.
 
 3.1.9 (2018-11-01)
 ==================